# docs/openapi.yaml
openapi: "3.0.0"
info:
  title: WhatsApp API MultiDevice
  version: 5.4.0
  description: This API is used for sending whatsapp via API
servers:
  - url: http://localhost:3000
...
security:
  - basicAuth: []
paths:
...
  /message/{message_id}/revoke:
    post:
      operationId: revokeMessage
      tags:
        - message
      summary: Revoke Message
      parameters:
        - in: path
          name: message_id
          schema:
            type: string
          required: true
          description: Message ID
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  example: '6289685024051@s.whatsapp.net'
                  description: Phone number with country code
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /message/{message_id}/delete:
    post:
      operationId: deleteMessage
      tags:
        - message
      summary: Delete Message
      parameters:
        - in: path
          name: message_id
          schema:
            type: string
          required: true
          description: Message ID
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  example: '6289685024051@s.whatsapp.net'
                  description: Phone number with country code
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /message/{message_id}/reaction:
    post:
      operationId: reactMessage
      tags:
        - message
      summary: Send reaction to message
      parameters:
        - in: path
          name: message_id
          schema:
            type: string
          required: true
          description: Message ID
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  example: '6289685024051@s.whatsapp.net'
                  description: Phone number with country code
                emoji:
                  type: string
                  example: "🙏"
                  description: Emoji to react
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /message/{message_id}/update:
    post:
      operationId: updateMessage
      tags:
        - message
      summary: Edit message by message ID before 15 minutes
      parameters:
        - in: path
          name: message_id
          schema:
            type: string
          required: true
          description: Message ID
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  example: '62819273192397132@s.whatsapp.net'
                  description: Phone number with country code
                message:
                  type: string
                  example: 'Hello World'
                  description: New message to send
              required:
                - phone
                - message
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /message/{message_id}/read:
    post:
      operationId: readMessage
      tags:
        - message
      summary: Mark as read message
      parameters:
        - in: path
          name: message_id
          schema:
            type: string
          required: true
          description: Message ID
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  example: '62819273192397132@s.whatsapp.net'
                  description: Phone number with country code
              required:
                - phone
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    SendResponse:
      type: object
      properties:
        code:
          type: string
          example: SUCCESS
        message:
          type: string
          example: Success
        results:
          type: object
          properties:
            message_id:
              type: string
              example: '3EB0B430B6F8F1D0E053AC120E0A9E5C'
            status:
              type: string
              example: '<feature> success ....'
...
    ErrorInternalServer:
      type: object
      properties:
        code:
          type: string
          example: INTERNAL_SERVER_ERROR
          description: 'SYSTEM_CODE_ERROR'
        message:
          type: string
          example: you are not loggin
          description: 'Detail error message'
        results:
          type: object
          example: null
          description: 'additional data'
    ErrorBadRequest:
      type: object
      properties:
        code:
          type: string
          example: 400
          description: 'HTTP Status Code'
        message:
          type: string
          example: field cannot be blank
          description: 'Detail error message'
        results:
          type: object
          example: null
          description: 'additional data'
...